requirements.txt
streamlit
mysql
scikit-learn
joblib
numpy
pandas
matplotlib
python-docx==1.1.0